inserted two flags in the init of the LC class for (i) keeping/removing the bk level, and (ii) use a Poisson/exponential distribution for p_2 and p_5
inserted a flag in the __init__ for choosing between an actual Poisson distribution or a (rounded) exponential when sampling the initial number of pulses and the number of child per pulses. Inserted a missing factor "self._ampl" in the function "_restore_lc()", in the expression for "self._plot_lc". Added a boolean flag for keeping or removing the background level at the end of the generation of each GRB.